# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Base Api",
        "description": "API do projeto Base",
        "contact": {
            "email": "email@test.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Base API"
        }
    ],
    "paths": {
...
        "/routes?totalPerPage={totalPerPage}&page={page}&filter={filter}": {
            "get": {
                "tags": [
                    "Permissão"
                ],
                "summary": "Listagem de rotas",
                "operationId": "listRotas",
                "parameters": [
                    {
                        "name": "totalPerPage",
                        "in": "path",
                        "description": "Total de items por página",
                        "required": false,
                        "allowEmptyValue": true,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    },
                    {
                        "name": "page",
                        "in": "path",
                        "description": "Número da página",
                        "required": false,
                        "allowEmptyValue": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "filter",
                        "in": "path",
                        "description": "Descrição do filtro",
                        "required": false,
                        "allowEmptyValue": true,
                        "schema": {
                            "type": "string",
                            "example": ""
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Listagem realizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Listagem realizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "O servidor não entende o tipo de conteúdo da entidade de solicitação",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Ocoreu um erro"
                    },
                    "404": {
                        "description": "Página não localizada"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/routes/{route}": {
            "get": {
                "tags": [
                    "Permissão"
                ],
                "summary": "Buscar Rota",
                "operationId": "buscarRota",
                "parameters": [
                    {
                        "name": "route",
                        "in": "path",
                        "description": "Id da rota",
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Busca realizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Busca realizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "O servidor não entende o tipo de conteúdo da entidade de solicitação",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Ocoreu um erro"
                    },
                    "404": {
                        "description": "Página não localizada"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/routes": {
            "post": {
                "tags": [
                    "Permissão"
                ],
                "summary": "Criar rota",
                "operationId": "storeRoute",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "test.store"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Registrar rota"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rota criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Rota criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "O servidor não entende o tipo de conteúdo da entidade de solicitação",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Ocoreu um erro"
                    },
                    "404": {
                        "description": "Página não localizada"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/permissions/{perfil}": {
            "post": {
                "tags": [
                    "Permissão"
                ],
                "summary": "Atribuir perfis a rota",
                "operationId": "storePermission",
                "parameters": [
                    {
                        "name": "perfil",
                        "in": "path",
                        "description": "ID da rota",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "routes"
                                ],
                                "properties": {
                                    "routes": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rota criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Rota criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "O servidor não entende o tipo de conteúdo da entidade de solicitação",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Ocoreu um erro"
                    },
                    "404": {
                        "description": "Página não localizada"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "in": "header",
                "flows": {
                    "clientCredentials": {
                        "tokenUrl": "/api/auth/auth",
                        "scopes": {}
                    }
                },
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}